Fix build with override map
The `OVERRIDE` definition was missing a semicolon and the value block was missing `map` as the return value. And the `Url::parse()` was lacking `.context()` call, not creating a proper `EasyError`. And `HashMap.get()` wants a key that can be borrowed so I converted the keys to `String`.

While at it, let’s also define the HashMap more declaratively, including the use of final types.