Use native rel=next links
With this, we no longer have to wait for the in-page next page links to be updated, which is often forgotten.

Since these links are automatically inserted by Wordpress, they include all blog posts.
But it should be fine, as Winkybubbles does not really post anything other than book contents.
In fact, we get stuff like e.g. the extra materials in Pale, which are not typically linked in-page.

Final chapters were marked as such in a previous commit, to avoid adding author’s retrospective or carry over from Worm to Glowworm.