Escape chapter titles
Some titles can contain XML-special characters, for example:

    https://palewebserial.wordpress.com/2020/05/28/1-6-spoilers-flyer-app/
    [1.6 Spoilers] Flyer & App

So we need to escape them when using the value in HTML.

Unfortunately, epub_builder currently does not escape the title in nav.xhtml,
only in toc.ncx, so we also need to modify the title passed to it.
Hopefully, e-book readers will be able to display the small ampersand.