Fix CloudFlare link replacement
Html5ever used internally by both select and scraper does not guarantee
attribute ordering when serializing HTML (using `inner_html()` function).
It was only luck that when the regex was implemented, the attribute map
mostly respected the order in the source code.

When we later started manipulating the code more heavily,
the unordered nature of Map became apparent,
giving us six different possible attribute sequences and,
as a result, the text has not been replaced.

Let’s use an actual DOM manipulation to avoid the wrath of Zalgo.